feat(charting): move validation function in controller
